Removed dependency on jQuery & updated manifest to version 2
I didn't like the idea of loading something as big as jQuery into my
page to get this done, so I wrote out the jQuery dependencies in the
extension and removed jQuery from the bundle, and from the manifest.

While I was at it, I updated the manifest to version 2.